+++ R/reconc_TDcond.R
@@ -44,6 +44,12 @@
                         toll=.TOLL, Rtoll=.RTOLL, smoothing=TRUE, 
                         al_smooth=.ALPHA_SMOOTHING, lap_smooth=.LAP_SMOOTHING) {
   
+  # If the bottom pmf list contains only 1 element,
+  # then the TD samples are simply a copy of the upper samples. 
+  if(length(bott_pmf)==1){
+    return(matrix(u, nrow=1))
+  }
+  
   l_l_pmf = rev(PMF.bottom_up(bott_pmf, toll = toll, Rtoll = Rtoll, return_all = TRUE, 
                               smoothing=smoothing, al_smooth=al_smooth, lap_smooth=lap_smooth))
   
@@ -88,7 +94,9 @@
 #' The base upper forecasts `fc_upper` must be a list containing the parameters of 
 #' the multivariate Gaussian distribution of the upper forecasts.
 #' The list must contain only the named elements `mu` (vector of length n_upper) 
-#' and `Sigma` (n_upper x n_upper matrix) 
+#' and `Sigma` (n_upper x n_upper matrix).
+#' 
+#' The order of the upper and bottom base forecasts must match the order of (respectively) the rows and the columns of A.
 #'  
 #' A PMF object is a numerical vector containing the probability mass function of a discrete distribution.
 #' Each element corresponds to the probability of the integers from 0 to the last value of the support.
@@ -99,7 +107,7 @@
 #' those samples are discarded and a warning is triggered.
 #' The warning reports the percentage of samples kept. 
 #'
-#' @param S Summing matrix (n x n_bottom).   
+#' @param A aggregation matrix (n_upper x n_bottom).   
 #' @param fc_bottom A list containing the bottom base forecasts, see details.
 #' @param fc_upper A list containing the upper base forecasts, see details.
 #' @param bottom_in_type A string with three possible values:
@@ -140,7 +148,6 @@
 #' 
 #' # Consider a simple hierarchy with two bottom and one upper
 #' A <- matrix(c(1,1),nrow=1)
-#' S <- rbind(A,diag(nrow=2))
 #' # The bottom forecasts are Poisson with lambda=15
 #' lambda <- 15
 #' n_tot <- 60
@@ -152,7 +159,7 @@
 #' fc_upper<- list(mu=40, Sigma=matrix(c(5^2)))
 #' 
 #' # We can reconcile with reconc_TDcond
-#' res.TDcond <- reconc_TDcond(S, fc_bottom, fc_upper)
+#' res.TDcond <- reconc_TDcond(A, fc_bottom, fc_upper)
 #' 
 #' # Note that the bottom distributions are shifted to the right
 #' PMF.summary(res.TDcond$bottom_reconciled$pmf[[1]])
@@ -165,17 +172,68 @@
 #' PMF.summary(res.TDcond$upper_reconciled$pmf[[1]])
 #' PMF.get_var(res.TDcond$upper_reconciled$pmf[[1]])
 #' 
+#' ## Example 2: reconciliation with unbalanced hierarchy
+#' # We consider the example in Fig. 9 of Zambon et al. (2024).
+#' 
+#' # The hierarchy has 5 bottoms and 3 uppers
+#' A <- matrix(c(1,1,1,1,1,
+#'               1,1,0,0,0,
+#'               0,0,1,1,0),nrow=3,byrow = TRUE)
+#' # Note that the 5th bottom only appears in the highest level, this is an unbalanced hierarchy. 
+#' n_upper  = nrow(A)
+#' n_bottom = ncol(A)
+#' 
+#' # The bottom forecasts are Poisson with lambda=15
+#' lambda <- 15
+#' n_tot <- 60
+#' fc_bottom <- list()
+#' for(i in seq(n_bottom)){
+#'   fc_bottom[[i]] <- apply(matrix(seq(0,n_tot)),MARGIN=1,FUN=function(x) dpois(x,lambda=lambda))
+#' }
+#'
+#' # The upper forecasts are a multivariate Gaussian
+#' mu = c(75, 30, 30)
+#' Sigma = matrix(c(5^2,5,5,
+#'                  5, 10, 0,
+#'                  5, 0,10), nrow=3, byrow = TRUE)
+#'                  
+#' fc_upper<- list(mu=mu, Sigma=Sigma)
+#' \dontrun{
+#' # If we reconcile with reconc_TDcond it won't work
+#' res.TDcond <- reconc_TDcond(A, fc_bottom, fc_upper)
+#' }
+#' 
+#' # We can balance the hierarchy with by duplicating the node b5
+#' # In practice this means: 
+#' # i) consider the time series observations for b5 as the upper u4,
+#' # ii) fit the multivariate ts model for u1, u2, u3, u4. 
+#' 
+#' # In this example we simply assume that the forecast for u1-u4 is 
+#' # Gaussian with the mean and variance of u4 given by the parameters in b5. 
+#' mean_b5 <- lambda
+#' var_b5  <- lambda
+#' mu = c(75, 30, 30,mean_b5)
+#' Sigma = matrix(c(5^2,5,5,5,
+#'                  5, 10, 0, 0,
+#'                  5, 0, 10, 0,
+#'                  5, 0,  0, var_b5), nrow=4, byrow = TRUE)
+#' fc_upper<- list(mu=mu, Sigma=Sigma)
+#' 
+#' # We also need to update the aggregating matrix
+#' A <- matrix(c(1,1,1,1,1,
+#'               1,1,0,0,0,
+#'               0,0,1,1,0,
+#'               0,0,0,0,1),nrow=4,byrow = TRUE)
+#'               
+#' # We can now reconcile with TDcond
+#' res.TDcond <- reconc_TDcond(A, fc_bottom, fc_upper)
+#' 
+#' # Note that the reconciled distribution of b5 and u4 are identical, 
+#' # keep this in mind when using the results of your reconciliation!
+#' max(abs(res.TDcond$bottom_reconciled$pmf[[5]]- res.TDcond$upper_reconciled$pmf[[4]]))
+#' 
+#' 
 #' @references
-#' Corani, G., Azzimonti, D., Augusto, J.P.S.C., Zaffalon, M. (2021). 
-#' *Probabilistic Reconciliation of Hierarchical Forecast via Bayes' Rule*. 
-#' ECML PKDD 2020. Lecture Notes in Computer Science, vol 12459.
-#' \doi{10.1007/978-3-030-67664-3_13}.
-#'
-#' Zambon, L., Agosto, A., Giudici, P., Corani, G. (2024). 
-#' *Properties of the reconciled distributions for Gaussian and count forecasts*. 
-#' International Journal of Forecasting (in press).
-#' \doi{10.1016/j.ijforecast.2023.12.004}.
-#'
 #' Zambon, L., Azzimonti, D., Rubattu, N., Corani, G. (2024). 
 #' *Probabilistic reconciliation of mixed-type hierarchical time series*. 
 #' The 40th Conference on Uncertainty in Artificial Intelligence, accepted.
@@ -183,7 +241,7 @@
 #' @seealso [reconc_MixCond()], [reconc_BUIS()]
 #'
 #' @export
-reconc_TDcond = function(S, fc_bottom, fc_upper, 
+reconc_TDcond = function(A, fc_bottom, fc_upper, 
                          bottom_in_type = "pmf", distr = NULL,
                          num_samples = 2e4, return_type = "pmf", 
                          suppress_warnings = FALSE, seed = NULL) {
@@ -191,16 +249,16 @@ reconc_TDcond = function(S, fc_bottom, fc_upper,
   if (!is.null(seed)) set.seed(seed)
   
   # Check inputs
-  .check_input_TD(S, fc_bottom, fc_upper, 
+  .check_input_TD(A, fc_bottom, fc_upper, 
                   bottom_in_type, distr,
                   return_type)
   
-  # Get aggr. matrix A and find the "lowest upper" 
-  A = .get_A_from_S(S)$A
+  # Find the "lowest upper" 
   n_u = nrow(A)
   n_b = ncol(A)
   lowest_rows = .lowest_lev(A)
   n_u_low = length(lowest_rows)  # number of lowest upper
+  n_u_upp = n_u - n_u_low        # number of "upper upper" 
   
   # Get mean and covariance matrix of the MVN upper base forecasts
   mu_u    = fc_upper$mu
@@ -216,14 +274,19 @@ reconc_TDcond = function(S, fc_bottom, fc_upper,
   } else {
     # Else, analytically reconcile the upper and then sample from the lowest-uppers
     
-    # Get the aggregation matrix A_u and the summing matrix S_u for the upper sub-hierarchy
+    # Get the aggregation matrix A_u for the upper sub-hierarchy
     A_u = .get_Au(A, lowest_rows)
-    S_u = matrix(nrow = n_u, ncol = n_u_low)
-    S_u[-lowest_rows,] = A_u
-    S_u[lowest_rows,] = diag(n_u_low)
     
     # Analytically reconcile the upper
-    rec_gauss_u = reconc_gaussian(S_u, mu_u, Sigma_u)
+    # The entries of mu_u must be in the correct order, i.e. rows of A_u (upper), columns of A_u (bottom)
+    mu_u_ord = c(mu_u[-lowest_rows],mu_u[lowest_rows])  
+    # Same for Sigma_u
+    Sigma_u_ord = matrix(nrow=n_u, ncol=n_u)
+    Sigma_u_ord[1:n_u_upp, 1:n_u_upp]             = Sigma_u[-lowest_rows,-lowest_rows]
+    Sigma_u_ord[1:n_u_upp, (n_u_upp+1):n_u]       = Sigma_u[-lowest_rows,lowest_rows]
+    Sigma_u_ord[(n_u_upp+1):n_u, 1:n_u_upp]       = Sigma_u[lowest_rows,-lowest_rows]
+    Sigma_u_ord[(n_u_upp+1):n_u, (n_u_upp+1):n_u] = Sigma_u[lowest_rows,lowest_rows]
+    rec_gauss_u = reconc_gaussian(A_u, mu_u_ord, Sigma_u_ord)
     
     # Sample from reconciled MVN on the lowest level of the upper (dim: num_samples x n_u_low)
     U = .MVN_sample(n_samples = num_samples,
@@ -254,7 +317,8 @@ reconc_TDcond = function(S, fc_bottom, fc_upper,
   samp_ok = rowSums(samp_ok) == n_u_low
   # Only keep the "good" upper samples, and throw a warning if some samples are discarded:
   U_js = lapply(U_js, "[", samp_ok) 
-  if (sum(samp_ok) != num_samples & !suppress_warnings) {
+  num_samples_ok = sum(samp_ok)
+  if (num_samples_ok != num_samples & !suppress_warnings) {
     # We round down to the nearest decimal
     warning(paste0("Only ", floor(sum(samp_ok)/num_samples*1000)/10, "% of the upper samples ",
                    "are in the support of the bottom-up distribution; ",
@@ -262,11 +326,11 @@ reconc_TDcond = function(S, fc_bottom, fc_upper,
   }
   
   # Get bottom samples via the prob top-down
-  B = list()
+  B = matrix(nrow = n_b, ncol = num_samples_ok)
   for (j in 1:n_u_low) {
-    B[[j]] = .TD_sampling(U_js[[j]], L_pmf_js[[j]])
+    mask_j = as.logical(A[lowest_rows[j], ])  # mask for the position of the bottom referring to lowest upper j
+    B[mask_j, ] = .TD_sampling(U_js[[j]], L_pmf_js[[j]])
   }
-  B = do.call("rbind", B)  # dim: n_bottom x num_samples
   U = A %*% B              # dim: n_upper x num_samples
   
   # Prepare output: include the marginal pmfs and/or the samples (depending on "return" inputs)
@@ -274,16 +338,12 @@ reconc_TDcond = function(S, fc_bottom, fc_upper,
   if (return_type %in% c('pmf', 'all')) {
     upper_pmf  = lapply(1:n_u, function(i) PMF.from_samples(U[i,]))
     bottom_pmf = lapply(1:n_b, function(i) PMF.from_samples(B[i,]))
-    
     out$bottom_reconciled$pmf = bottom_pmf
     out$upper_reconciled$pmf = upper_pmf
-    
   }
   if (return_type %in% c('samples','all')) {
-    
     out$bottom_reconciled$samples = B
     out$upper_reconciled$samples = U
-    
   } 
 
   return(out)