+++ tests/testthat/test-weightswarn.R
@@ -12,14 +12,15 @@ test_that("Test effective sample size", {
   
   w = .compute_weights(b, u, "samples", "continuous")
   
-  check_w = .check_weigths(w, n_eff_min=200)
+  check_w = .check_weights(w, n_eff_min=200)
   expect_equal(check_w$warning, TRUE)
   expect_equal(check_w$warning_code, 1)
   expect_equal(check_w$n_eff, n)
   
   # Try the warning message
   # base_forecast = list(u,b1,b2)
-  # a = reconc_BUIS(S, base_forecast, in_type = "samples", distr = list("continuous","discrete","discrete"), seed=42)
+  # A = matrix(S[1,],nrow=1)
+  # a = reconc_BUIS(A, base_forecast, in_type = "samples", distr = list("continuous","discrete","discrete"), seed=42)
   
   # -----------
   n = 199
@@ -32,14 +33,15 @@ test_that("Test effective sample size", {
   
   w = .compute_weights(b, u, "samples", "continuous")
   
-  check_w = .check_weigths(w, n_eff_min=200)
+  check_w = .check_weights(w, n_eff_min=200)
   expect_equal(check_w$warning, TRUE)
   expect_equal(check_w$warning_code, 1)
   expect_equal(check_w$n_eff, n)
   
   # Try the warning message
   # base_forecast = list(u,b1,b2)
-  # a = bayesRecon::reconc_BUIS(S, base_forecast, in_type = "samples", distr = list("continuous","discrete","discrete"), seed=42)
+  # A = matrix(S[1,],nrow=1)
+  # a = bayesRecon::reconc_BUIS(A, base_forecast, in_type = "samples", distr = list("continuous","discrete","discrete"), seed=42)
 
   # -----------
   n = 2000
@@ -52,12 +54,13 @@ test_that("Test effective sample size", {
   
   w = .compute_weights(b, u, "samples", "continuous")
   
-  check_w = .check_weigths(w, n_eff_min=200, p_n_eff=0.01)
+  check_w = .check_weights(w, n_eff_min=200, p_n_eff=0.01)
   expect_equal(check_w$warning, TRUE)
   expect_equal(check_w$warning_code, c(2,3))
   expect_equal(check_w$n_eff < 200, TRUE)
   
   # Try the warning message
   # base_forecast = list(u,b1,b2)
-  # a = bayesRecon::reconc_BUIS(S, base_forecast, in_type = "samples", distr = list("continuous","discrete","discrete"), seed=42)
+  # A = matrix(S[1,],nrow=1)
+  # a = bayesRecon::reconc_BUIS(A, base_forecast, in_type = "samples", distr = list("continuous","discrete","discrete"), seed=42)
 })